+++ src/ModuarkServiceProvider.php
@@ -30,7 +30,9 @@ use Cluion\Moduark\Console\ModuleGraphCommand;
 use Cluion\Moduark\Console\ModuleListCommand;
 use Cluion\Moduark\Discovery\ModuleDiscoverer;
 use Cluion\Moduark\Graph\CapabilityGraphBuilder;
+use Cluion\Moduark\Graph\Export\MermaidCapabilityGraphExporter;
 use Cluion\Moduark\Graph\Export\MermaidModuleGraphExporter;
+use Cluion\Moduark\Graph\Export\TextCapabilityGraphExporter;
 use Cluion\Moduark\Graph\Export\TextModuleGraphExporter;
 use Cluion\Moduark\Graph\ModuleGraphBuilder;
 use Cluion\Moduark\Lifecycle\ModuleLifecycleRegistrar;
@@ -86,6 +88,8 @@ final class ModuarkServiceProvider extends ServiceProvider
         $this->app->singleton(SourceIndexBuilder::class);
         $this->app->singleton(CapabilityGraphBuilder::class);
         $this->app->singleton(ModuleGraphBuilder::class);
+        $this->app->singleton(TextCapabilityGraphExporter::class);
+        $this->app->singleton(MermaidCapabilityGraphExporter::class);
         $this->app->singleton(TextModuleGraphExporter::class);
         $this->app->singleton(MermaidModuleGraphExporter::class);
         $this->app->singleton(ModuleOrderer::class);